AlphaCC: Add Feature - Add AlphaCC code and remove GCC dependency
1. Add AlphaCC code
2. Remove GCC code from the repo
