+++ BUILD.gn
@@ -527,7 +527,6 @@ if (rtc_include_tests) {
   rtc_test("slow_tests") {
     testonly = true
     deps = [
-      "modules/congestion_controller/goog_cc:goog_cc_slow_tests",
       "rtc_base/task_utils:repeating_task_unittests",
       "test:test_main",
     ]
