+++ api/transport/BUILD.gn
@@ -69,17 +69,18 @@ rtc_source_set("field_trial_based_config") {
   ]
 }
 
-rtc_static_library("goog_cc") {
+# Revision for enabling AlphaCC and disabling GCC
+rtc_static_library("alpha_cc") {
   visibility = [ "*" ]
   sources = [
-    "goog_cc_factory.cc",
-    "goog_cc_factory.h",
+    "alpha_cc_factory.cc",
+    "alpha_cc_factory.h",
   ]
   deps = [
     ":network_control",
     ":webrtc_key_value_config",
     "..:network_state_predictor_api",
-    "../../modules/congestion_controller/goog_cc",
+    "../../modules/congestion_controller/alpha_cc",
     "//third_party/abseil-cpp/absl/memory",
   ]
 }
