+++ call/rtp_transport_controller_send.cc
@@ -12,7 +12,8 @@
 
 #include "absl/memory/memory.h"
 #include "absl/types/optional.h"
-#include "api/transport/goog_cc_factory.h"
+// Revision for enabling AlphaCC and disabling GCC
+#include "api/transport/alpha_cc_factory.h"
 #include "api/transport/network_types.h"
 #include "api/units/data_rate.h"
 #include "api/units/time_delta.h"
