+++ modules/congestion_controller/alpha_cc/BUILD.gn
@@ -1,0 +1,41 @@
+# Copyright (c) 2018 The WebRTC project authors. All Rights Reserved.
+#
+# Use of this source code is governed by a BSD-style license
+# that can be found in the LICENSE file in the root of the source
+# tree. An additional intellectual property rights grant can be found
+# in the file PATENTS.  All contributing project authors may
+# be found in the AUTHORS file in the root of the source tree.
+
+import("../../../webrtc.gni")
+
+config("bwe_test_logging") {
+  if (rtc_enable_bwe_test_logging) {
+    defines = [ "BWE_TEST_LOGGING_COMPILE_TIME_ENABLE=1" ]
+  } else {
+    defines = [ "BWE_TEST_LOGGING_COMPILE_TIME_ENABLE=0" ]
+  }
+}
+
+rtc_static_library("alpha_cc") {
+  configs += [ ":bwe_test_logging" ]
+  sources = [
+    "alpha_cc_network_control.cc",
+    "alpha_cc_network_control.h",
+  ]
+
+  # There is no dependency now
+  deps = [
+  ]
+}
+
+rtc_source_set("link_capacity_estimator") {
+  sources = [
+    "link_capacity_estimator.cc",
+    "link_capacity_estimator.h",
+  ]
+  deps = [
+    "../../../api/units:data_rate",
+    "../../../rtc_base:safe_minmax",
+    "//third_party/abseil-cpp/absl/types:optional",
+  ]
+}
