+++ modules/congestion_controller/alpha_cc/link_capacity_estimator.cc
@@ -7,7 +7,7 @@
  *  in the file PATENTS.  All contributing project authors may
  *  be found in the AUTHORS file in the root of the source tree.
  */
-#include "modules/congestion_controller/goog_cc/link_capacity_estimator.h"
+#include "modules/congestion_controller/alpha_cc/link_capacity_estimator.h"
 
 #include <algorithm>
