+++ modules/remote_bitrate_estimator/BUILD.gn
@@ -49,7 +49,8 @@ rtc_static_library("remote_bitrate_estimator") {
     "../../api/units:timestamp",
     "../../modules:module_api",
     "../../modules:module_api_public",
-    "../../modules/congestion_controller/goog_cc:link_capacity_estimator",
+    # Revision for enabling AlphaCC and disabling GCC
+    "../../modules/congestion_controller/alpha_cc:link_capacity_estimator",    
     "../../modules/rtp_rtcp:rtp_rtcp_format",
     "../../rtc_base:checks",
     "../../rtc_base:rtc_base_approved",
