+++ test/scenario/BUILD.gn
@@ -114,7 +114,9 @@ if (rtc_include_tests) {
       "../../modules/audio_device:mock_audio_device",
       "../../modules/audio_mixer:audio_mixer_impl",
       "../../modules/audio_processing",
-      "../../modules/congestion_controller/goog_cc:test_goog_cc_printer",
+      # Remove it for enabling AlphaCC and disabling GCC
+      # Todo: The test doesn't work now      
+      # "../../modules/congestion_controller/goog_cc:test_goog_cc_printer",
       "../../modules/rtp_rtcp",
       "../../modules/rtp_rtcp:mock_rtp_rtcp",
       "../../modules/rtp_rtcp:rtp_rtcp_format",
